Changed LineTime Day Displaying as int not string of day Changed Time to a custom double representation of time of day.

+++ SOFA/Models/LineTime.cs
@@ -13,7 +13,7 @@ namespace SOFA.Models
 
         public Int32 Day { get; set; }
 
-        public Int32 Time { get; set; }
+        public Double Time { get; set; }
 
         public virtual Line Line { get; set; }
     }
